+++ R/app_ui.R
@@ -8,7 +8,13 @@ app_ui <- function() {
     qmongr::add_external_resources(),
     # List the first level UI elements here
     shiny::tagList(
-      quality_overview_ui("quality_overview_ui_1")
+      quality_overview_ui("quality_overview_ui_1"),
+      shiny::tags$script(src = "www/d3.min.js"),
+      shiny::tags$script(src = "www/theme.js"),
+      shiny::tags$script(src = "www/axis.js"),
+      shiny::tags$script(src = "www/line_chart.js"),
+      shiny::tags$script(src = "www/barchart.js"),
+      shiny::tags$script(src = "www/qmongr.js")
     )
   )
 }