+++ R/quality_overview.R
@@ -40,6 +40,9 @@ quality_overview_ui <- function(id) {
             outputId = ns("qi_table")
           )
         )
+      ),
+      shiny::htmlOutput(
+        ns("json")
       )
     )
   )
@@ -124,5 +127,17 @@ quality_overview_server <- function(id) {
   output$qi_overview <- shiny::renderUI({
     sidebar_qo_ui("quality_overview_ui_1")
   })
+
+  #data passed to js
+  output$json <- shiny::reactive({
+      paste("<script> var  description = ",
+            jsonlite::toJSON(app_data$register_data$description, na = "null"), ";",
+            "var indicator_hosp =", jsonlite::toJSON(app_data$grouped_by_hospital), ";",
+            "var indicator_hf =", jsonlite::toJSON(app_data$grouped_by_hf), ";",
+            "var indicator_rhf =", jsonlite::toJSON(app_data$grouped_by_rhf), ";",
+            "var indicator_nat =", jsonlite::toJSON(app_data$national_data), ";",
+            "; </script>")
+  })
+
   })
 }